OneDrive/Pictures/document_reasoning_assistant/requirements.txt
fastapi
uvicorn
python-dotenv
torch
transformers
sentence-transformers
faiss-cpu
pymupdf
python-docx
psycopg2-binary
requests
pinecone
